Optimized for VCT v2.0 and enhanced plots
The VCT v2.0 creates a reliable polyfit when cal_opt.step_thres is set to  0.4 mT.

Better visibility for detecting polyfit-errors in the plots.
Please check if the amount of countet steps (in red) is = cal_opt.small_step_range + 1 (e.g. range from -24 to 24 degrees with a step size of 2 deg should result in 25 steps).